fix: made version just a script

+++ version.ts
@@ -16,38 +16,35 @@ export const IMPORT_MAP_REGEX = globToRegExp("examples/**/importMap.json", {
   caseInsensitive: false,
 });
 
-/** `prepublish` will be invoked before publish, return `false` to prevent the publish. */
-export async function prepublish(version: string) {
-  const newDenoLandVersion = `//deno.land/x/ultra@v${version}/`;
+if (import.meta.main) {
+  const version = prompt("Whats the new version?");
+  if (version) {
+    const newDenoLandVersion = `//deno.land/x/ultra@v${version}/`;
 
-  const readme = await Deno.readTextFile("./README.md");
+    const readme = await Deno.readTextFile("./README.md");
 
-  console.log("Patching README.md");
-  await Deno.writeTextFile(
-    "./README.md",
-    readme.replace(
-      VERSION_REGEX,
-      `//deno.land/x/ultra@v${version}/`,
-    ),
-  );
-
-  console.log("Patching examples importMaps");
-
-  for await (
-    const entry of walk("./", {
-      match: [IMPORT_MAP_REGEX],
-      skip: [ULTRA_OUTPUT_REGEX],
-    })
-  ) {
-    const content = await Deno.readTextFile(entry.path);
+    console.log("Patching README.md");
     await Deno.writeTextFile(
-      entry.path,
-      content.replace(VERSION_REGEX, newDenoLandVersion),
+      "./README.md",
+      readme.replace(
+        VERSION_REGEX,
+        `//deno.land/x/ultra@v${version}/`,
+      ),
     );
-  }
-}
 
-/** `postpublish` will be invoked after published. */
-export function postpublish(version: string) {
-  console.log("Upgraded to", version);
+    console.log("Patching examples importMaps");
+
+    for await (
+      const entry of walk("./", {
+        match: [IMPORT_MAP_REGEX],
+        skip: [ULTRA_OUTPUT_REGEX],
+      })
+    ) {
+      const content = await Deno.readTextFile(entry.path);
+      await Deno.writeTextFile(
+        entry.path,
+        content.replace(VERSION_REGEX, newDenoLandVersion),
+      );
+    }
+  }
 }
